Removed the in-progress files from master
Also discovered a slight bug with the directory listener, which is that post removal fires more than once for the same post.

+++ app/helpers/posts_helper.rb
@@ -287,7 +287,8 @@ module PostsHelper
     return unless post.exists?
     post = post.first
 
-    post.destroy
+    # guarding against multiple fires
+    post.destroy unless post.nil?
   end
 
   # load all posts that we may not have in database
